New data collector for debugging, metrics, backup & restore and others
It's allow us to collect data easily and allow we're format it to JSON, XML, YAML, TOML, INI, Markdown even Excel sheets and HTML!

So we can use it to create a more human-readable web gui for debugging and others and it more stable and readable than currently we using stupid plain-text paster.

Also it contains more datas so it will more powerful and allow us use script analytics the possibly bugs and conflicts with other plugins.

And it will be a part of QuickShop CrashHandler System


Former-commit-id: 5200130eeb9a9dbd83d7e5b84818204ccb17565b